Tests/LibWeb/Text/expected/aia-cert-fetching.txt

DER intermediate via AIA: loaded
PKCS#7 intermediate via AIA: loaded
dead AIA URL: failed
untrusted intermediate (must not validate): failed
both hosts sharing an intermediate: loaded, loaded
